Character Behavior
This is the current project that I'm working on. It is an RTS/MOBA game
and I'm responsible for networking the game and setting up the behaviour
and animations for different characters.

Pokemon is the base class from which all the characters derive from.
This base class has an update function which gets called every frame.
This update function in turn calls the update the for the current state
that the character is present in.

Machoke_Walk is one of many states that derive from States class. This
base class has OnEnter(), Update() and OnExit() which is overridden by
different states to suit their behaviour.

ClientGameObjectMangerAddon contains the different events that gets
passed from one client to another client.
Each character has a ghost ID which is unique across all clients. Also,
each character has a view of the network(m_networkView) which stores its
own GhostID and other info about the network. A network manager then
stores all the network views in the game.

When a particular client receives an event, it checks the ghostID and
modified that particular character in the game.