Program that returns the list of words given a boggle table and dictionary
I came across this particular problem statement on the internet and  I wanted to try it out. I also had read about "trie" data structure but didn't have a chance to implement it. As soon as I read the problem, I knew that trie would be very useful in this case.

Given the table size and list of characters in the table, this program returns the list of words that can be formed using the given letters and are valid according to the given dictionary as well. 

The words are entered into the trie as individual letters and the last letter of a word is marked. 
The logic which optimizes this program is "Search for a word based on prefixes. If a prefix is not present, then the word will not be present as well." Program starts with the first letter of the table and recursively adds the neighbouring letter to the current word. For Example, if  the current word is "pre" so far, then it would check if p-r-e is present in the trie. If its not, it terminates. If it is present, then it takes the letter neighbouring to 'e' and adds it to the current word and checks again and so on.